RSPY-664 - use smaller nodes for staging
Staging does not need a lot of cpu or ram, but aims at dispatch several
dask workers on several nodes to maximize the network throughput.

Our tests show that when scaling the workers on b3-16 instances, several
dask workers are hosted on a same node, this sharing the same network
interface. Testing with Matera and Svalbard, we got network speed
between 300 Mb/s to 600 Mb/s on a single worker, we never managed to get
higher speed, close to the 1 Gb/s limit advertised for b3-16 instance.

By using a smaller node model, we maximize the chance to have the dask
workers split on several nodes, thus increasing the overall speed, at a
lower cost.